requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
email-validator==2.2.0
anthropic==0.84.0
reportlab==4.4.10
psycopg2-binary==2.9.10
